Fix bug when running setLoc in InputText
When running setLoc in InputText (after initialization), the fields: imageRect, rect, focusedImageRect and cursorLoc are still using the old loc.
This causes a weird behaviour where you need to click the previous location to be able to edit the input field and the animations also happen in the wrong place.